+++ test-sql/test.sql
@@ -25,7 +25,7 @@ begin
     select 1=1
 end
 
--- #region section debug 
+-- #region section debug (4 lines inside)
 begin
     select 1=1
     select 1=1